+++ app/src/main/res/layout/activity_parsing.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ParsingActivity">
+    tools:context=".ui.ParsingActivity">
  
 
 </androidx.constraintlayout.widget.ConstraintLayout>
